Complete Phase 5 filter teardown
