+++ src/filters/filterstate.cpp
@@ -17,9 +17,7 @@ FilterState MakeDefaultState(const FilterSpec &spec)
     return std::visit(
         [](const auto &payload) -> FilterState {
             using Payload = std::decay_t<decltype(payload)>;
-            if constexpr (std::is_same_v<Payload, LegacyPayload>) {
-                return LegacyState{};
-            } else if constexpr (std::is_same_v<Payload, TextPayload>) {
+            if constexpr (std::is_same_v<Payload, TextPayload>) {
                 return TextState{};
             } else if constexpr (std::is_same_v<Payload, ComboPayload>) {
                 return ComboState{};
