+++ src/search.cpp
@@ -9,17 +9,13 @@
 #include "bucket.h"
 #include "buyoutmanager.h"
 #include "column.h"
-#include "filters.h"
 #include "filters/filtermatchers.h"
 #include "filters/filterspec.h"
 #include "items_model.h"
 #include "util/fatalerror.h"
 #include "util/spdlog_qt.h" // IWYU pragma: keep
 
-Search::Search(BuyoutManager &bo_manager,
-               const QString &caption,
-               const FilterCatalog &catalog,
-               const std::vector<Filter *> &filters)
+Search::Search(BuyoutManager &bo_manager, const QString &caption, const FilterCatalog &catalog)
     : m_bo_manager(bo_manager)
     , m_filter_catalog(catalog)
     , m_model(bo_manager, *this)
@@ -29,8 +25,6 @@ Search::Search(BuyoutManager &bo_manager,
     , m_current_mode(ViewMode::ByTab)
     , m_refresh_reason(RefreshReason::Unknown)
 {
-    Q_ASSERT(filters.size() == static_cast<size_t>(m_filter_catalog.size()));
-
     using move_only = std::unique_ptr<Column>;
     move_only init[] = {std::make_unique<NameColumn>(),
                         std::make_unique<PriceColumn>(bo_manager),
@@ -61,62 +55,44 @@ Search::Search(BuyoutManager &bo_manager,
     m_columns = std::vector<move_only>(std::make_move_iterator(std::begin(init)),
                                        std::make_move_iterator(std::end(init)));
 
+    m_filter_states.reserve(static_cast<size_t>(m_filter_catalog.size()));
     for (qsizetype index = 0; index < m_filter_catalog.size(); ++index) {
         const auto &spec = m_filter_catalog[index];
-        if (std::holds_alternative<LegacyPayload>(spec.payload)) {
-            auto *filter = filters.at(static_cast<size_t>(index));
-            Q_ASSERT(filter);
-            m_filter_slots.emplace_back(filter->CreateData());
-        } else {
-            FilterState state = MakeDefaultState(spec);
-            const bool matchingState = std::visit(
-                [&state](const auto &payload) {
-                    using Payload = std::decay_t<decltype(payload)>;
-                    if constexpr (std::is_same_v<Payload, TextPayload>) {
-                        return std::holds_alternative<TextState>(state);
-                    } else if constexpr (std::is_same_v<Payload, ComboPayload>) {
-                        return std::holds_alternative<ComboState>(state);
-                    } else if constexpr (std::is_same_v<Payload, MinMaxPayload>) {
-                        return std::holds_alternative<MinMaxState>(state);
-                    } else if constexpr (std::is_same_v<Payload, ColorsPayload>) {
-                        return std::holds_alternative<ColorsState>(state);
-                    } else if constexpr (std::is_same_v<Payload, BoolPayload>) {
-                        return std::holds_alternative<BoolState>(state);
-                    } else if constexpr (std::is_same_v<Payload, ModsPayload>) {
-                        return std::holds_alternative<ModsState>(state);
-                    } else {
-                        return false;
-                    }
-                },
-                spec.payload);
-            Q_ASSERT(matchingState);
-            m_filter_slots.emplace_back(std::move(state));
-        }
+        FilterState state = MakeDefaultState(spec);
+        const bool matchingState = std::visit(
+            [&state](const auto &payload) {
+                using Payload = std::decay_t<decltype(payload)>;
+                if constexpr (std::is_same_v<Payload, TextPayload>) {
+                    return std::holds_alternative<TextState>(state);
+                } else if constexpr (std::is_same_v<Payload, ComboPayload>) {
+                    return std::holds_alternative<ComboState>(state);
+                } else if constexpr (std::is_same_v<Payload, MinMaxPayload>) {
+                    return std::holds_alternative<MinMaxState>(state);
+                } else if constexpr (std::is_same_v<Payload, ColorsPayload>) {
+                    return std::holds_alternative<ColorsState>(state);
+                } else if constexpr (std::is_same_v<Payload, BoolPayload>) {
+                    return std::holds_alternative<BoolState>(state);
+                } else {
+                    return std::holds_alternative<ModsState>(state);
+                }
+            },
+            spec.payload);
+        Q_ASSERT(matchingState);
+        m_filter_states.emplace_back(std::move(state));
     }
+    Q_ASSERT(m_filter_states.size() == static_cast<size_t>(m_filter_catalog.size()));
 }
 
 Search::~Search() = default;
 
-FilterData &Search::legacyFilterDataAt(qsizetype index)
-{
-    auto *filter = std::get_if<std::unique_ptr<FilterData>>(
-        &m_filter_slots.at(static_cast<size_t>(index)));
-    Q_ASSERT(filter && *filter);
-    return **filter;
-}
-
 FilterState &Search::filterStateAt(qsizetype index)
 {
-    auto *state = std::get_if<FilterState>(&m_filter_slots.at(static_cast<size_t>(index)));
-    Q_ASSERT(state);
-    return *state;
+    return m_filter_states.at(static_cast<size_t>(index));
 }
 
 const FilterState &Search::filterStateAt(qsizetype index) const
 {
-    const auto *state = std::get_if<FilterState>(&m_filter_slots.at(static_cast<size_t>(index)));
-    Q_ASSERT(state);
-    return *state;
+    return m_filter_states.at(static_cast<size_t>(index));
 }
 
 void Search::setExpandedHeaders(std::set<QString> headers)
@@ -234,20 +210,10 @@ void Search::FilterItems(const Items &items)
     // active, so we don't have to check every filter against
     // every item.
     std::vector<qsizetype> active_filters;
-    active_filters.reserve(m_filter_slots.size());
-    for (qsizetype index = 0; index < static_cast<qsizetype>(m_filter_slots.size()); ++index) {
-        const auto &slot = m_filter_slots.at(static_cast<size_t>(index));
-        if (const auto *filter = std::get_if<std::unique_ptr<FilterData>>(&slot)) {
-            Q_ASSERT(*filter);
-            if ((*filter)->filter()->IsActive()) {
-                active_filters.push_back(index);
-            }
-        } else {
-            const auto *state = std::get_if<FilterState>(&slot);
-            Q_ASSERT(state);
-            if (state && IsActive(*state)) {
-                active_filters.push_back(index);
-            }
+    active_filters.reserve(m_filter_states.size());
+    for (qsizetype index = 0; index < static_cast<qsizetype>(m_filter_states.size()); ++index) {
+        if (IsActive(m_filter_states.at(static_cast<size_t>(index)))) {
+            active_filters.push_back(index);
         }
     }
     active_filters.shrink_to_fit();
@@ -272,17 +238,8 @@ void Search::FilterItems(const Items &items)
         // current item.
         bool matches = true;
         for (const qsizetype index : active_filters) {
-            const auto &slot = m_filter_slots.at(static_cast<size_t>(index));
-            bool filterMatches = false;
-            if (const auto *filter = std::get_if<std::unique_ptr<FilterData>>(&slot)) {
-                Q_ASSERT(*filter);
-                filterMatches = (*filter)->Matches(item);
-            } else {
-                const auto *state = std::get_if<FilterState>(&slot);
-                Q_ASSERT(state);
-                filterMatches = state && MatchesFilter(*item, m_filter_catalog[index], *state);
-            }
-            if (!filterMatches) {
+            const auto &state = m_filter_states.at(static_cast<size_t>(index));
+            if (!MatchesFilter(*item, m_filter_catalog[index], state)) {
                 // Now that we know this item will be filtered out,
                 // we don't need to check any more filters.
                 matches = false;
