+++ src/ui/mainwindow.cpp
@@ -870,10 +870,7 @@ void MainWindow::NewSearch()
     m_tab_bar->addTab("+");
 
     spdlog::trace("MainWindow::NewSearch() setting current search: {}", caption);
-    m_current_search = new Search(m_buyout_manager,
-                                  caption,
-                                  m_filter_catalog,
-                                  m_search_form->legacyFilters());
+    m_current_search = new Search(m_buyout_manager, caption, m_filter_catalog);
     m_current_search->SetRefreshReason(RefreshReason::TabCreated);
 
     // this can't be done in ctor because it'll call OnSearchFormChange slot
