+++ tests/CMakeLists.txt
@@ -17,5 +17,6 @@ acq_add_test(tst_buyoutmanager)
 acq_add_test(tst_itemtooltiptext)
 acq_add_test(tst_itemsmodel)
 acq_add_test(tst_search)
+acq_add_test(tst_searchform)
 acq_add_test(tst_filters)
 acq_add_test(tst_workerparse)
